% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadtext.R
\name{loadtext}
\alias{loadtext}
\title{Loading text file}
\usage{
loadtext()
}
\arguments{
\item{Example.txt}{This is an example text file}
}
\value{
An object called textfile is created in the global environment.
}
\description{
Loading text file
}
\examples{
loadtext()
}
