% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{ratio}
\alias{ratio}
\title{Calculating ratio between positive and negative words}
\usage{
ratio(pos, neg)
}
\arguments{
\item{negative_count}{A vector of counts of negative words found in a text}

\item{positive_count}{A vector of counts of positive words found in a text}
}
\value{
The ratio
}
\description{
Calculating ratio between positive and negative words
}
\examples{
ratio(positive_count, negative_count)
}
