% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_words.R
\name{scoring_words}
\alias{scoring_words}
\title{Scoring a vector of text based on words}
\usage{
scoring_words(words, text = textfile)
}
\arguments{
\item{words}{A list of words you want to search for/score. Use * to use prefixes or suffixes.}

\item{textfile}{The text file you want to search in, e.g. textfile.}
}
\value{
A vector of length(text) with scores for each word.
}
\description{
Scoring a vector of text based on words
}
\examples{
negative_words <- c("overwhelming", "*ar", "trapped", "frustrating", "a", "co*", "*ate")
positive_words <- c("shake", "cooperate", "woke", "diary", "bed", "i", "sle*")
scoring_words(negative_words)
}
