% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_words.R
\name{match_words}
\alias{match_words}
\title{match words}
\usage{
match_words(words, lexicon)
}
\arguments{
\item{words}{A character vector of words from the text.}

\item{lexicon}{A vector with a lexicon.}
}
\value{
A vector of matched words found in the text.
}
\description{
Match words in text using preloaded positive and negative lexicons

This function compares a vector of words from a text to the internal
Positive and Negative vectors. It supports "*" as a wildcard suffix
or prefix.
}
\examples{
words <- c("happy", "joyful", "nauseous", "nauseated")
lexicon <- c("happ*", "joy*")
match_words(words, lexicon)
}
