% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_words.R
\name{match_words}
\alias{match_words}
\title{match words}
\usage{
match_words(words, type = c("positive", "negative"))
}
\arguments{
\item{words}{A character vector of words from the text.}

\item{type}{Either "positive" or "negative" to select the lexicon.}
}
\value{
A vector of matched words found in the text.
}
\description{
Match words in text using preloaded positive and negative lexicons

This function compares a vector of words from a text to the internal
Positive and Negative vectors. It supports "*" as a wildcard suffix
or prefix.
}
\examples{
words <- c("happy", "joyful", "nauseous", "nauseated")
match_words(words, type = "positive")
match_words(words, type = "negative")
}
